requirements.txt
streamlit==1.31.0
pandas==2.2.0
google-generativeai==0.3.2
google-genai
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6